requirements.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.0.4
django-filter==21.1
django-tables2==2.4.1
djangorestframework==3.13.1
executing==0.8.3
importlib-metadata==4.11.3
ipython==8.3.0
jedi==0.18.1
Markdown==3.3.7
matplotlib-inline==0.1.3
parso==0.8.3
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
python-dateutil==2.8.2
pytz==2022.1
pytzdata==2020.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
zipp==3.8.0
